requirements.txt
langchain
langchain-core
langchain-openai
langchain-community
langchain-pinecone
langgraph
streamlit
python-dotenv
tiktoken